Move settings.py to habitat_sim.utils
