+++ tests/test_data_extraction.py
@@ -11,10 +11,10 @@ from torch.nn import functional as F
 from torch.utils.data import DataLoader, Dataset
 
 import habitat_sim
-from examples.settings import make_cfg
 from habitat_sim.utils.data.data_extractor import ImageExtractor
 from habitat_sim.utils.data.data_structures import ExtractorLRUCache
 from habitat_sim.utils.data.pose_extractor import TopdownView
+from habitat_sim.utils.settings import make_cfg
 
 
 class TrivialNet(nn.Module):
