+++ tests/test_physics_benchmarking.py
@@ -6,8 +6,8 @@
 
 import pytest
 
-import examples.settings
 import habitat_sim
+import habitat_sim.utils.settings
 import utils
 
 
@@ -40,10 +40,10 @@ def test_benchmark_urdf_add_remove(benchmark, iterations, force_reload):
     # test loading and removing a URDF ArticultedObject multiple times consecutively
     def instance_remove_urdf(iterations):
         # first configure the simulator
-        cfg_settings = examples.settings.default_sim_settings.copy()
+        cfg_settings = habitat_sim.utils.settings.default_sim_settings.copy()
         cfg_settings["scene"] = "NONE"
         cfg_settings["enable_physics"] = True
-        hab_cfg = examples.settings.make_cfg(cfg_settings)
+        hab_cfg = habitat_sim.utils.settings.make_cfg(cfg_settings)
         with habitat_sim.Simulator(hab_cfg) as sim:
             art_obj_mgr = sim.get_articulated_object_manager()
