+++ tests/test_random_seed.py
@@ -1,11 +1,11 @@
-import examples.settings
 import habitat_sim
+import habitat_sim.utils.settings
 
 
 def test_random_seed():
     # reconfigure to ensure pathfinder exists
-    cfg_settings = examples.settings.default_sim_settings.copy()
-    hab_cfg = examples.settings.make_cfg(cfg_settings)
+    cfg_settings = habitat_sim.utils.settings.default_sim_settings.copy()
+    hab_cfg = habitat_sim.utils.settings.make_cfg(cfg_settings)
     with habitat_sim.Simulator(hab_cfg) as sim:
         # Test that the same seed gives the same point
         sim.seed(1)
